PYIC-1558: Rename authorizationcode lambda to be check-passport.
