+++ settings.gradle
@@ -3,7 +3,7 @@ include "lib",
 		"lambdas",
 		"lambdas:accesstoken",
 		"lambdas:issuecredential",
-		"lambdas:authorizationcode",
+		"lambdas:checkpassport",
 		"lambdas:buildclientoauthresponse",
 		"lambdas:initialisesession",
 		"integration-test"
